added

+++ USL/exam-01/03a.c
@@ -1,21 +1,21 @@
+// Write a C program to remove empty files from the given directory.
+
 #include <stdio.h>
 #include <fcntl.h>
 #include <unistd.h>
 #include <dirent.h>
+
 int main(){
     DIR *dp;
     struct dirent *dir;
-    int fd,n;
+    dp = opendir("."); 
 
-    dp = opendir(".");
-
-    if(dp){
-            while((dir = readdir(dp)) != NULL){
-                    fd = open(dir->d_name,O_RDWR,0777);
-                    n = lseek(fd,0,SEEK_END);
-                    if(!n){
-                            unlink(dir->d_name);
-                    }
-            }
+    while ((dir = readdir(dp)) != NULL){
+        int fd = open(dir->d_name, O_RDWR, 0777);
+        int n = lseek(fd, 0, SEEK_END);
+        
+        if (!n) unlink(dir->d_name);
     }
+    
+    return 0;
 }
